src/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
lunar-python==1.3.10
pydantic>=2.7.4
pydantic-settings>=2.10.1
httpx>=0.27.0
langchain-openai==1.1.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
fastapi-limiter==0.1.6
redis==5.0.1
